DRY ContactPresenter and WorldwideOrganisation::LinkedContactPresenter
The `contact` schema is used for both contact pages, as well as the contacts
related to WorldwideOffices and WorldwideOrganisations.

However, the ContactPresenter and WorldwideOrganisation::LinkedContactPresenter
are quite separate in their implementations.

This adds a module to DRY up any common code.

We may want to consider using the AddressFormatter for Contacts as a further
refinement.
