+++ app/presenters/contact_presenter.rb
@@ -1,21 +1,11 @@
 class ContactPresenter < ContentItemPresenter
   include ContentItem::TitleAndContext
+  include ContentItem::ContactDetails
 
   def title_and_context
     super.tap do |t|
       t.delete(:average_title_length)
       t.delete(:context)
-    end
-  end
-
-  def online_form_links
-    contact_form_links = content_item["details"]["contact_form_links"] || []
-    contact_form_links.map do |link|
-      {
-        url: link["link"],
-        title: link["title"],
-        description: link["description"].try(:html_safe),
-      }
     end
   end
 
@@ -125,18 +115,6 @@ private
       v_card_class:,
       value: value.try(:strip).try(:html_safe),
     }
-  end
-
-  def email_address_groups
-    content_item["details"]["email_addresses"] || []
-  end
-
-  def post_address_groups
-    content_item["details"]["post_addresses"] || []
-  end
-
-  def phone_number_groups
-    content_item["details"]["phone_numbers"] || []
   end
 
   def related_contacts_links
